fix: BoardDef pin number invalid type, #51

+++ src/Main/BoardDef.h
@@ -76,7 +76,7 @@ struct BoardDef {
 
   /** Watchdog */
   struct {
-    const uint8_t resetPin;
+    const int resetPin;
     const bool supported;
   } WDG;
   const char *name;
